+++ FSIC/__init__.py
@@ -12,4 +12,4 @@ models in the tradition of Wynne Godley.
 from FSIC.version import MAJOR, MINOR, PATCH, DEV, VERSION as __version__  # noqa
 from FSIC.settings import DTYPE  # noqa
 
-from FSIC.classes.schematic import Variable
+from FSIC.classes.schematic import Variable, Equation